Kablammo/graph.js merged with kablammo.js
